requirements.txt
pytest>=7.4.3
requests>=2.26.0
selenium>=4.1.0
allure-pytest>=2.13.2